Revert "this should have been long..."
This reverts commit 23b3930ab9651f8520e8cad38f72646d81167b67.

+++ src/com/roman/romcontrol/fragments/Weather.java
@@ -59,7 +59,7 @@ public class Weather extends SettingsPreferenceFragment implements OnPreferenceC
 
         mWeatherSyncInterval = (ListPreference) findPreference("refresh_interval");
         mWeatherSyncInterval.setOnPreferenceChangeListener(this);
-        mWeatherSyncInterval.setSummary(Long.toString(WeatherPrefs.getRefreshInterval(mContext))
+        mWeatherSyncInterval.setSummary(Integer.toString(WeatherPrefs.getRefreshInterval(mContext))
                 + " minutes");
 
         mCustomWeatherLoc = (EditTextPreference) findPreference("custom_location");
@@ -175,7 +175,7 @@ public class Weather extends SettingsPreferenceFragment implements OnPreferenceC
     @Override
     public boolean onPreferenceChange(Preference preference, Object newValue) {
         if (preference == mWeatherSyncInterval) {
-            long newVal = Long.parseLong((String) newValue);
+            int newVal = Integer.parseInt((String) newValue);
             preference.setSummary(newValue + " minutes");
 
             return WeatherPrefs.setRefreshInterval(mContext, newVal);
